01-Predictions/requirements.txt
sklearn
pandas
pyarrow
imblearn
pandas-gbq
google-cloud-bigquery
google-cloud-storage
google-auth-oauthlib